move: build -> root
