+++ config.js
@@ -2,7 +2,7 @@ const path = require('path');
 const { argv } = require('yargs');
 const merge = require('webpack-merge');
 
-const userConfig = require('../src/config');
+const userConfig = require('./src/config.json');
 const isProduction = !!((argv.env && argv.env.production) || argv.p);
 const rootPath = process.cwd();
