+++ package.json
@@ -17,9 +17,9 @@
     "theme"
   ],
   "scripts": {
-    "build": "webpack --mode development --progress --config build/webpack.config.js",
-    "build:production": "webpack --mode production --progress -p --config build/webpack.config.js",
-    "watch": "webpack --hide-modules --watch --config build/webpack.config.js",
+    "build": "webpack --mode development --progress --config webpack.config.js",
+    "build:production": "webpack --mode production --progress -p --config webpack.config.js",
+    "watch": "webpack --hide-modules --watch --config webpack.config.js",
     "lint:js": "eslint --ext .js,.vue src/scripts/* src/components/*",
     "lint:sass": "sass-lint -c .sass-lint.yml -q -v",
     "format:js": "eslint --fix --ext .js,.vue src/scripts/* src/components/*",
@@ -34,9 +34,10 @@
       "git add"
     ]
   },
-  "engines": {
-    "node": ">= 8.0.0"
-  },
+  "browserslist": [
+    "ie >= 11",
+    "safari >= 10"
+  ],
   "devDependencies": {
     "@babel/core": "7.0.0-beta.54",
     "@babel/plugin-proposal-object-rest-spread": "7.0.0-beta.54",
