+++ src/config.json
@@ -1,7 +1,3 @@
 {
-  "publicPath": "/wp-content/themes/b.0218.jp",
-  "browsers": [
-    "ie >= 11",
-    "safari >= 10"
-  ]
+  "publicPath": "/wp-content/themes/b.0218.jp"
 }
